# - componetize content of the main page
