+++ src/App.css
@@ -3,6 +3,7 @@
   margin: 0 auto;
   padding: 2rem;
   text-align: center;
+  display: 'flex',
 }
 
 .logo {
@@ -40,3 +41,5 @@
 .read-the-docs {
   color: #888;
 }
+
+
