+++ src/components/Footer.tsx
@@ -1,0 +1,9 @@
+const FooterSection = () => {
+    return (
+        <div style={{ textAlign: 'center' }}>
+            Task Manager ©2025 Created by KemJiga and JonesProgrammer
+        </div>
+    );
+}
+
+export default FooterSection;
