+++ src/components/Header.tsx
@@ -1,0 +1,14 @@
+import type { MenuProps } from 'antd';
+
+const items1: MenuProps['items'] = [
+    {
+        key: 'viewButton',
+        label: 'view',
+    },
+    {
+        key: 'reportButton',
+        label: 'report',
+    },
+];
+
+export default items1;
