+++ src/pages/HomePage.tsx
@@ -1,234 +1,15 @@
 import React from 'react';
-import dayjs from 'dayjs';
-import 'dayjs/locale/zh-cn';
-import type { BadgeProps } from 'antd';
-import { Badge } from 'antd';
-import { Calendar, Col, Radio, Row, Select, Typography } from 'antd';
-import type { CalendarProps } from 'antd';
-import type { Dayjs } from 'dayjs';
-import dayLocaleData from 'dayjs/plugin/localeData';
-import { UserOutlined, BookOutlined, PlusOutlined, CalendarOutlined } from '@ant-design/icons';
-import type { MenuProps } from 'antd';
+import { PlusOutlined, UserOutlined } from '@ant-design/icons';
 import { Layout, Menu, Button, theme } from 'antd';
-
-dayjs.extend(dayLocaleData);
+import CustomCalendar from '../components/CustomCalendar';
+import items1 from '../components/Header';
+import items2 from '../components/Aside';
+import { Footer } from 'antd/es/layout/layout';
+import FooterSection from '../components/Footer';
 
 const { Header, Content, Sider } = Layout;
 
-const getListData = (value: Dayjs) => {
-  let listData: { type: string; content: string }[] = [];
-  switch (value.date()) {
-    case 8:
-      listData = [
-        { type: 'warning', content: 'This is warning event.' },
-        { type: 'success', content: 'This is usual event.' },
-      ];
-      break;
-    case 10:
-      listData = [
-        { type: 'warning', content: 'This is warning event.' },
-        { type: 'success', content: 'This is usual event.' },
-        { type: 'error', content: 'This is error event.' },
-      ];
-      break;
-    case 15:
-      listData = [
-        { type: 'warning', content: 'This is warning event' },
-        { type: 'success', content: 'This is very long usual event......' },
-        { type: 'error', content: 'This is error event 1.' },
-        { type: 'error', content: 'This is error event 2.' },
-        { type: 'error', content: 'This is error event 3.' },
-        { type: 'error', content: 'This is error event 4.' },
-      ];
-      break;
-    default:
-  }
-  return listData || [];
-};
-
-const getMonthData = (value: Dayjs) => {
-  if (value.month() === 8) {
-    return 1394;
-  }
-};
-
-const CustomCalendar: React.FC = () => {
-  const monthCellRender = (value: Dayjs) => {
-    const num = getMonthData(value);
-    return num ? (
-      <div className="notes-month">
-        <section>{num}</section>
-        <span>Backlog number</span>
-      </div>
-    ) : null;
-  };
-
-  const dateCellRender = (value: Dayjs) => {
-    const listData = getListData(value);
-    return (
-      <ul className="events">
-        {listData.map((item) => (
-          <li key={item.content}>
-            <Badge status={item.type as BadgeProps['status']} text={item.content} />
-          </li>
-        ))}
-      </ul>
-    );
-  };
-
-  const cellRender: CalendarProps<Dayjs>['cellRender'] = (current, info) => {
-    if (info.type === 'date') return dateCellRender(current);
-    if (info.type === 'month') return monthCellRender(current);
-    return info.originNode;
-  };
-
-  return <Calendar cellRender={cellRender} />;
-};
-
-const CustomminiCalendar: React.FC = () => {
-  const { token } = theme.useToken();
-
-  const onPanelChange = (value: Dayjs, mode: CalendarProps<Dayjs>['mode']) => {
-    console.log(value.format('YYYY-MM-DD'), mode);
-  };
-
-  const wrapperStyle: React.CSSProperties = {
-    width: '100%',
-    border: `1px solid ${token.colorBorderSecondary}`,
-    borderRadius: token.borderRadiusLG,
-    padding: '10px',
-  };
-
-  return (
-    <div style={wrapperStyle}>
-      <Calendar
-        fullscreen={false}
-        headerRender={({ value, type, onChange, onTypeChange }) => {
-          const start = 0;
-          const end = 12;
-          const monthOptions = [];
-
-          let current = value.clone();
-          const localeData = value.localeData();
-          const months = [];
-          for (let i = 0; i < 12; i++) {
-            current = current.month(i);
-            months.push(localeData.monthsShort(current));
-          }
-
-          for (let i = start; i < end; i++) {
-            monthOptions.push(
-              <Select.Option key={i} value={i} className="month-item">
-                {months[i]}
-              </Select.Option>
-            );
-          }
-
-          const year = value.year();
-          const month = value.month();
-          const options = [];
-          for (let i = year - 10; i < year + 10; i += 1) {
-            options.push(
-              <Select.Option key={i} value={i} className="year-item">
-                {i}
-              </Select.Option>
-            );
-          }
-          return (
-            <div style={{ padding: 8 }}>
-              <Typography.Title level={4}><CalendarOutlined /> {value.format('YYYY-MM')}</Typography.Title>
-              <Row gutter={8}>
-                <Col>
-                  <Radio.Group
-                    size="small"
-                    onChange={(e) => onTypeChange(e.target.value)}
-                    value={type}
-                  >
-                    <Radio.Button value="month">Month</Radio.Button>
-                    <Radio.Button value="year">Year</Radio.Button>
-                  </Radio.Group>
-                </Col>
-                <Col>
-                  <Select
-                    size="small"
-                    popupMatchSelectWidth={false}
-                    className="my-year-select"
-                    value={year}
-                    onChange={(newYear) => {
-                      const now = value.clone().year(newYear);
-                      onChange(now);
-                    }}
-                  >
-                    {options}
-                  </Select>
-                </Col>
-                <Col>
-                  <Select
-                    size="small"
-                    popupMatchSelectWidth={false}
-                    value={month}
-                    onChange={(newMonth) => {
-                      const now = value.clone().month(newMonth);
-                      onChange(now);
-                    }}
-                  >
-                    {monthOptions}
-                  </Select>
-                </Col>
-              </Row>
-            </div>
-          );
-        }}
-        onPanelChange={onPanelChange}
-      />
-    </div>
-  );
-};
-
-const items1: MenuProps['items'] = [
-  {
-    key: 'leftButton',
-    label: (
-      <Button type="primary" style={{ marginRight: '0px' }}>
-        <UserOutlined />
-      </Button>
-    ),
-  },
-  {
-    key: 'viewButton',
-    label: 'view',
-  },
-  {
-    key: 'reportButton',
-    label: 'report',
-  },
-];
-
-const items2: MenuProps['items'] = [
-  {
-    key: 'calendar',
-    label: <CustomminiCalendar />,
-    type: 'group',
-  },
-  {
-    key: 'activities',
-    label: (
-      <div style={{ padding: '10px' }}>
-        <div style={{ fontWeight: 'bold', marginBottom: '8px' }}>
-          <BookOutlined /> Actividades
-        </div>
-        <ul style={{ listStyleType: 'none', padding: 0, margin: 0 }}>
-          <li style={{ padding: '5px 0' }}>Actividad 1</li>
-          <li style={{ padding: '5px 0' }}>Actividad 2</li>
-          <li style={{ padding: '5px 0' }}>Actividad 3</li>
-        </ul>
-      </div>
-    ),
-    type: 'group',
-  },
-];
-
-const App: React.FC = () => {
+const HomePage: React.FC = () => {
   const {
     token: { colorBgContainer, borderRadiusLG },
   } = theme.useToken();
@@ -237,6 +18,9 @@ const App: React.FC = () => {
     <Layout>
       <Header style={{ display: 'flex', alignItems: 'center' }}>
         <div className="demo-logo" />
+        <Button type="primary" style={{ marginRight: '20px' }}>
+          <UserOutlined />
+        </Button>
         <Menu
           theme="dark"
           mode="horizontal"
@@ -249,7 +33,7 @@ const App: React.FC = () => {
       </Header>
 
       <Layout>
-        <Sider width={200} style={{ background: colorBgContainer }}>
+        <Sider width={350} style={{ background: colorBgContainer }}>
           <Menu
             mode="inline"
             defaultSelectedKeys={['1']}
@@ -272,8 +56,13 @@ const App: React.FC = () => {
           </Content>
         </Layout>
       </Layout>
+      <Layout>
+        <Footer>
+          <FooterSection />
+        </Footer>
+      </Layout>
     </Layout>
   );
 };
 
-export default App;
+export default HomePage;
